+++ wagtail_wordpress_import/importers/wordpress.py
@@ -8,13 +8,12 @@ from django.apps import apps
 from django.utils.text import slugify
 from django.utils.timezone import make_aware
 from wagtail.core.models import Page
-from wagtail_wordpress_import.bleach import bleach_clean, fix_styles
+from wagtail_wordpress_import.bleach import bleach_clean
 from wagtail_wordpress_import.block_builder import BlockBuilder
-from wagtail_wordpress_import.functions import (
-    linebreaks_wp,
-    node_to_dict,
-    normalize_style_attrs,
-)
+from wagtail_wordpress_import.fix_styles import fix_styles
+from wagtail_wordpress_import.functions import node_to_dict
+from wagtail_wordpress_import.linebreaks_wp import linebreaks_wp
+from wagtail_wordpress_import.normalize_styles import normalize_style_attrs
 
 
 class WordpressImporter:
